Fix compiling warnings on Linux

+++ sample/sample_benchmark_std_couroutine_callable_recursive.cpp
@@ -44,7 +44,7 @@ benchmark_callable_future_type run_benchmark(size_t idx, int left_recursive_coun
     auto gen_res = co_await *first_hang;
     result += gen_res;
   } else {
-    result += idx;
+    result += static_cast<int64_t>(idx);
   }
 
   if (left_recursive_count > 1) {
